Clarifies useEffect API Documentation.
I really only understood this by referencing the react-reason code itself.  I understood that `2` to mean "this is the second version of the thing we're talking about" not "call this specially named function"  (obvious in retrospect, but....).

It still took me a while to realize that I was expected to return a cleanup function from useEffect even if I didn't need to do cleanup (something I'm not used to having to do).  But I guess that begs the question how detailed ya'll want your docs to be here.